Refactor logics
