+++ src/constants/fullcalendar/settings.ts
@@ -1,0 +1,29 @@
+// field name
+export const FIELD_NAME = {
+  H1: "FIELD__H1",
+  H2: "FIELD__H2",
+} as const;
+
+// keys
+export const GROUP_ID_KEY = "GROUP_ID_KEY";
+export const RESOURCE_NAME_KEY = "RESOURCE_NAME_KEY";
+
+// resouce ID
+export const RESOURCE_ID__SHARED__AGE = "RESOURCE_ID__SHARED__AGE";
+export const RESOURCE_ID__SHARED__LIMIT = "RESOURCE_ID__SHARED__LIMIT";
+export const RESOURCE_ID__VISA__STUDY = "RESOURCE_ID__VISA__STUDY";
+export const RESOURCE_ID__VISA__COOP = "RESOURCE_ID__VISA__COOP";
+export const RESOURCE_ID__VISA__WORKING_HOLIDAY =
+  "RESOURCE_ID__VISA__WORKING_HOLIDAY";
+export const RESOURCE_ID__STATUS__WORKER = "RESOURCE_ID__STATUS__WORKER";
+export const RESOURCE_ID__STATUS__STUDENT = "RESOURCE_ID__STATUS__STUDENT";
+export const RESOURCE_ID__ETC__CUSTOM = "RESOURCE_ID__ETC__CUSTOM";
+export type RESOURCE_ID =
+  | typeof RESOURCE_ID__SHARED__AGE
+  | typeof RESOURCE_ID__SHARED__LIMIT
+  | typeof RESOURCE_ID__VISA__STUDY
+  | typeof RESOURCE_ID__VISA__COOP
+  | typeof RESOURCE_ID__VISA__WORKING_HOLIDAY
+  | typeof RESOURCE_ID__STATUS__WORKER
+  | typeof RESOURCE_ID__STATUS__STUDENT
+  | typeof RESOURCE_ID__ETC__CUSTOM;
